Extract UIImage test extension to UIImage+TestHelpers file
